# Optimize Enum.concat/1 for lists of lists