@greg-rychlewski 2021-06-21
It seems like the compiler doesn't like `&is_list/1` here:

```
if is_list(enumerables) and :lists.all(&is_list/1, enumerables) do
```

If I change it to this the warning disappears

```
if is_list(enumerables) and :lists.all(fn x -> is_list(x) end, enumerables) do
```