@v0idpwn 2021-06-21
Hey! I think making ConcatTest a protocol would give us more precise benchmark results. While your approach is of course still faster, I think it would give us better history for posteriority :)