@greg-rychlewski 2021-06-21
> Hey! I think making ConcatTest a protocol would give us more precise benchmark results. While your approach is of course still faster, I think it would give us better history for posteriority :)

Thanks! I switched it to a protocol and re-ran the tests. Here are the new results:

```
Operating System: macOS
CPU Information: Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz
Number of Available Cores: 8
Available memory: 16 GB
Elixir 1.9.4
Erlang 22.0

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
parallel: 1
inputs: large, medium, small
Estimated total run time: 42 s

Benchmarking master with input large...
Benchmarking master with input medium...
Benchmarking master with input small...
Benchmarking proposal with input large...
Benchmarking proposal with input medium...
Benchmarking proposal with input small...

##### With input large #####
Name               ips        average  deviation         median         99th %
proposal        817.79        1.22 ms    ±18.28%        1.28 ms        1.73 ms
master          349.81        2.86 ms    ±42.11%        2.52 ms        7.88 ms

Comparison: 
proposal        817.79
master          349.81 - 2.34x slower +1.64 ms

##### With input medium #####
Name               ips        average  deviation         median         99th %
proposal       13.34 K       74.96 μs    ±38.21%          70 μs         173 μs
master          5.97 K      167.59 μs    ±24.77%         141 μs      311.67 μs

Comparison: 
proposal       13.34 K
master          5.97 K - 2.24x slower +92.63 μs

##### With input small #####
Name               ips        average  deviation         median         99th %
proposal        1.14 M        0.87 μs  ±3032.45%           1 μs           1 μs
master          0.51 M        1.95 μs  ±1278.05%           2 μs           3 μs

Comparison: 
proposal        1.14 M
master          0.51 M - 2.23x slower +1.07 μs
```