@josevalim 2021-06-21
Thanks @greg-rychlewski!

I pushed a small change that inlines :lists.append implementation, so it doesn't require us to check the whole list upfront. It also has the benefit of using the fast approach until we find an element that is not a list. Your implementation would also return an improper list if the last list in concat was improper, so the change also address it.